[#77]: Create helpers.hpp for renderer.cpp local functions
